申请VIP，更多个人消息和查看对方更多消息的界面和功能 修改关于年龄的界面和数据，

+++ CommunityProject/CommunityProject/Activity/C/PlatFormActController.m
@@ -136,7 +136,7 @@
     PlatformActListModel * model = self.dataArr[indexPath.row];
     UIStoryboard * sb = [UIStoryboard storyboardWithName:@"Activity" bundle:nil];
     PlatformDetailController * detail = [sb instantiateViewControllerWithIdentifier:@"PlatformDetailController"];
-    detail.idStr = [NSString stringWithFormat:@"%ld",model.id];
+    detail.idStr = [NSString stringWithFormat:@"%ld",(long)model.id];
     [self.navigationController pushViewController:detail animated:YES];
 
 }
