+++ CommunityProject/CommunityProject/Address/V/MaybeKnowCell.m
@@ -37,7 +37,7 @@
     [self.headImageView sd_setImageWithURL:[NSURL URLWithString:encodeUrl]];
     self.nameLabel.text = _friendModel.nickname;
     self.numberLabel.text = _friendModel.numberId;
-    self.commonLabel.text = [NSString stringWithFormat:@"%ld位共同好友",_friendModel.count];
+    self.commonLabel.text = [NSString stringWithFormat:@"%ld位共同好友",(long)_friendModel.count];
     
     switch (_friendModel.relationship) {
         case 0:
